pii_detector/requirements.txt
requests

django
djangorestframework
django-cors-headers

pii-codex
nltk
pii-codex[detections]

openai
